feat: реализована модульность
